requirements.txt
openai 
tiktoken 
langchain
langchain-experimental
python-dotenv
rich
pydantic
Flask
requests